usual save

+++ C++/make_things_paralel/mander/mandelrajz.hpp
@@ -1,6 +1,6 @@
 #ifndef MANDELRAJZ_HPP_INCLUDED
 #define MANDELRAJZ_HPP_INCLUDED
 
-void mandelrajz(int XX, int YY, int MAX,float x_,float y_,float mag) ;
+void mandelrajz(int XX, int YY, int MAX,double x_,double y_,double mag) ;
 
 #endif // MANDELRAJZ_HPP_INCLUDED
